Add aws-cli and kubectl
This took a little doing due to Alpine's weird-musl behavior.

Take authorized_keys from the building user's
.ssh directory.

Add tmux to the installed packages. Trim the Dockerfile a bit, update
and edit the README. --privileged is needed for the docker run
command, that let me run zerotier in the pod on my mac.