A workable shell image
With the development tools I need for work stuff. Along with some
advice on running Zerotier within it.

Generally I'm going to fire up the pod on a k8s cluster and attach it
to a ZeroTier network in order to get regular access to it.

This thing will also form a perfectly viable image for running Node
and Go executables in production while still allowing for direct
interactive access to the pod itself.