Make Emacs work
Removed the yaml file in favor of using the `kubectl debug` command to
fire up the pod as this comes with pretty much all that stuff plus a
few things to allow for better management of nodes.

Emacs in the Oracle repos is pretty old. So, build our own!

Of course there are a couple dependencies (help2man and texinfo) that
are also not in any repos so build them also.

Emacs is a fairly big thing and during the make process segfaults
started happening. So I cut the make process into a few steps and have
it retry a bunch of times in the main make process (that was failing
so bad). This replaces the simple `make install` which was pretty much
exactly the same as the other two installs.

The build command that was failing was:

    docker buildx build --platform linux/amd64 --build-arg AUTH_KEYS="$(base64 -i .ssh/authorized_keys)" --build-arg USER=allen -t allenluce/shell .

Notable is the --platform flag which I needed to use as I was building
this on Apple Silicon (aarch64). Without this flag the build had no
hiccups at all. It'd be nice at some point to figure out what the
segfaults are due to, fix that directly, and simplify the emacs
installation command.

But for the time being, this does the job.